Now using location popup config coming from the config file to display location popup UI elements
