+++ extra/api-playground/api-config.json
@@ -26,7 +26,12 @@
             "text": "Demo Application",
             "link": "https://demos.location.aws.dev/"
           }
-        ]
+        ],
+        "locationPopupConfig": {
+          "showPlaceId": true,
+          "showLatitude": true,
+          "showLongitude": true
+        }
       }
     ],
     "Geocode": [
@@ -55,7 +60,12 @@
             "text": "Demo Application",
             "link": "https://demos.location.aws.dev/"
           }
-        ]
+        ],
+        "locationPopupConfig": {
+          "showPlaceId": false,
+          "showLatitude": true,
+          "showLongitude": true
+        }
       }
     ]
   }
