+++ src/atomicui/molecules/LocationPopup/index.ts
@@ -1,4 +1,4 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. */
 /* SPDX-License-Identifier: MIT-0 */
 
-export { default as Popup } from "./Popup";
+export { default as LocationPopup } from "./LocationPopup";
