+++ src/atomicui/molecules/LocationPopup/styles.scss
