+++ src/atomicui/molecules/index.ts
@@ -2,7 +2,7 @@
 /* SPDX-License-Identifier: MIT-0 */
 
 export * from "./MapButtons";
-export * from "./Popup";
+export * from "./LocationPopup";
 export * from "./ApiCard";
 export * from "./AutoCompleteLatLonInput";
 export * from "./CheckboxGroup";
@@ -10,5 +10,5 @@ export * from "./FilterModal";
 export * from "./MapButtons";
 export * from "./MapMarker";
 export * from "./Modal";
-export * from "./Popup";
+export * from "./LocationPopup";
 export * from "./SliderWithInput";
